refactor(wallet-mobile): network config -> manager
