+++ apps/wallet-mobile/src/features/WalletManager/common/helpers/date-to-epoch-info.test.ts
@@ -1,9 +1,9 @@
-import {byronEraConfig, networksConfig, shelleyEraConfig} from '../networks'
+import {byronEraConfig, networkManager, shelleyEraConfig} from '../network-manager'
 import {dateToEpochInfo} from './date-to-epoch-info'
 
 describe('dateToEpochInfo', () => {
   it('should return the correct epoch information', () => {
-    const convertDateToEpoch = dateToEpochInfo(networksConfig['mainnet'].eras)
+    const convertDateToEpoch = dateToEpochInfo(networkManager['mainnet'].eras)
 
     const inputDate = new Date('2024-05-14T12:30:00Z')
     const expectedOutput = {
@@ -19,7 +19,7 @@ describe('dateToEpochInfo', () => {
   })
 
   it('should throw an error for a date before the start of the known eras', () => {
-    const convertDateToEpoch = dateToEpochInfo(networksConfig['mainnet'].eras)
+    const convertDateToEpoch = dateToEpochInfo(networkManager['mainnet'].eras)
 
     const inputDate = new Date('1998-12-31T23:59:59Z')
 
@@ -27,7 +27,7 @@ describe('dateToEpochInfo', () => {
   })
 
   it('should return the first byron epoch - 0', () => {
-    const convertDateToEpoch = dateToEpochInfo(networksConfig['mainnet'].eras)
+    const convertDateToEpoch = dateToEpochInfo(networkManager['mainnet'].eras)
 
     const inputDate = new Date(1506203091000)
     const expectedOutput = {
@@ -43,7 +43,7 @@ describe('dateToEpochInfo', () => {
   })
 
   it('should return the last byron epoch - 207', () => {
-    const convertDateToEpoch = dateToEpochInfo(networksConfig['mainnet'].eras)
+    const convertDateToEpoch = dateToEpochInfo(networkManager['mainnet'].eras)
 
     const inputDate = new Date(Date.UTC(2020, 6, 29, 21, 44, 50))
     const expectedOutput = {
@@ -59,7 +59,7 @@ describe('dateToEpochInfo', () => {
   })
 
   it('should return the first shelley epoch - 208', () => {
-    const convertDateToEpoch = dateToEpochInfo(networksConfig['mainnet'].eras)
+    const convertDateToEpoch = dateToEpochInfo(networkManager['mainnet'].eras)
 
     const inputDate = new Date(Date.UTC(2020, 6, 29, 21, 44, 51))
