+++ apps/wallet-mobile/src/features/WalletManager/common/types.ts
@@ -1,13 +1,13 @@
 import {Portfolio} from '@yoroi/types'
 
-export type NetworkConfig = {
+export type NetworkManager = {
   primaryTokenInfo: Portfolio.Token.Info
   chainId: number
   eras: ReadonlyArray<NetworkEraConfig>
 }
 
 export type NetworkEraConfig = {
-  name: 'byron' | 'shelley' | 'conway'
+  name: 'byron' | 'shelley'
   start: Date
   end: Date | undefined
   slotInSeconds: number
@@ -19,4 +19,15 @@ export type NetworkEpochInfo = {
   start: Date
   end: Date
   era: NetworkEraConfig
+}
+
+export type NetworkEpochProgress = {
+  progress: number
+  currentSlot: number
+  timeRemaining: {
+    days: number
+    hours: number
+    minutes: number
+    seconds: number
+  }
 }
